response to ben comments
